--- a/gradle/libs.versions.toml
+++ b/gradle/libs.versions.toml
@@ -14,7 +14,7 @@ gradle-publish-maven = "0.29.0"
 
 kotlincrypto-core = "0.5.3"
 kotlincrypto-endians = "0.3.1"
-kotlincrypto-sponges = "0.3.1"
+kotlincrypto-sponges = "0.3.2"
 
 [libraries]
 gradle-kmp-configuration = { module = "io.matthewnelson:gradle-kmp-configuration-plugin", version.ref = "gradle-kmp-configuration" }