`kotlincrypto.core` -> `0.5.4`
`kotlincrypto.sponges` -> `0.3.2`