@05nelsonm 2024-12-19
Comparison with updated `sponges` and `core` dependencies with their performance improvements. 

<details>
    <summary>macOS</summary>

Before:
```
jvm summary:
Benchmark                        Mode  Cnt      Score       Error  Units
M55Benchmark.digest              avgt    5    571.578 ±     9.801  ns/op
ParallelHash128Benchmark.digest  avgt    5  80142.902 ± 10846.328  ns/op
SHA1Benchmark.digest             avgt    5    646.447 ±     7.691  ns/op
SHA256Benchmark.digest           avgt    5    811.096 ±    68.459  ns/op
SHA3_256Benchmark.digest         avgt    5  20934.448 ±  [251](https://github.com/KotlinCrypto/hash/actions/runs/12413539131/job/34655637827#step:5:252)0.715  ns/op
SHA512Benchmark.digest           avgt    5   1126.758 ±    11.224  ns/op
SHAKE128Benchmark.digest         avgt    5  23902.835 ±  5452.423  ns/op
TupleHash128Benchmark.digest     avgt    5  32843.986 ±  9388.236  ns/op

js summary:
Benchmark                        Mode  Cnt       Score       Error  Units
M55Benchmark.digest              avgt    5    1244.977 ±    22.843  ns/op
SHA1Benchmark.digest             avgt    5    1642.189 ±    81.793  ns/op
SHA256Benchmark.digest           avgt    5    1818.762 ±    79.350  ns/op
SHA512Benchmark.digest           avgt    5  116260.824 ±   495.920  ns/op
ParallelHash128Benchmark.digest  avgt    5  376199.935 ±  6906.651  ns/op
SHA3_256Benchmark.digest         avgt    5  119894.299 ±  1612.095  ns/op
SHAKE128Benchmark.digest         avgt    5  126031.375 ± 13546.923  ns/op
TupleHash128Benchmark.digest     avgt    5  16[312](https://github.com/KotlinCrypto/hash/actions/runs/12413539131/job/34655639073#step:5:313)7.407 ±  2261.124  ns/op

wasmJs summary:
Benchmark                        Mode  Cnt       Score       Error  Units
M55Benchmark.digest              avgt    5     781.930 ±    16.504  ns/op
SHA1Benchmark.digest             avgt    5    1105.501 ±    12.149  ns/op
SHA256Benchmark.digest           avgt    5    1398.059 ±    27.867  ns/op
SHA512Benchmark.digest           avgt    5    2531.700 ±    11.215  ns/op
ParallelHash128Benchmark.digest  avgt    5  185266.824 ± 12072.370  ns/op
SHA3_256Benchmark.digest         avgt    5   57455.400 ±  1125.052  ns/op
SHAKE128Benchmark.digest         avgt    5   57487.315 ±   606.947  ns/op
TupleHash128Benchmark.digest     avgt    5   77277.403 ±   657.163  ns/op

nativeHost summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5     497.725 ±    6.290  ns/op
ParallelHash128Benchmark.digest  avgt    5  344998.816 ± 5331.683  ns/op
SHA1Benchmark.digest             avgt    5     854.916 ±    1.316  ns/op
SHA256Benchmark.digest           avgt    5    1152.926 ±    6.155  ns/op
SHA3_256Benchmark.digest         avgt    5  112287.776 ±  463.266  ns/op
SHA512Benchmark.digest           avgt    5    1553.345 ±    1.179  ns/op
SHAKE128Benchmark.digest         avgt    5  116073.254 ±  469.291  ns/op
TupleHash128Benchmark.digest     avgt    5  157596.220 ± 2240.925  ns/op
```

After:
```
jvm summary:
Benchmark                        Mode  Cnt      Score      Error  Units
M55Benchmark.digest              avgt    5    553.324 ±    2.495  ns/op
ParallelHash128Benchmark.digest  avgt    5  60773.248 ± 3480.816  ns/op
SHA1Benchmark.digest             avgt    5    663.346 ±   42.127  ns/op
SHA256Benchmark.digest           avgt    5    762.972 ±   18.541  ns/op
SHA3_256Benchmark.digest         avgt    5  13944.237 ±  185.730  ns/op
SHA512Benchmark.digest           avgt    5   1120.699 ±   26.008  ns/op
SHAKE128Benchmark.digest         avgt    5  14097.538 ±  253.123  ns/op
TupleHash128Benchmark.digest     avgt    5  20178.017 ± 2231.180  ns/op

js summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5    1595.938 ±  206.210  ns/op
SHA1Benchmark.digest             avgt    5    1792.849 ±  380.365  ns/op
SHA256Benchmark.digest           avgt    5    1782.593 ±   23.869  ns/op
SHA512Benchmark.digest           avgt    5  114093.258 ±  601.467  ns/op
ParallelHash128Benchmark.digest  avgt    5  292545.426 ± 3311.475  ns/op
SHA3_256Benchmark.digest         avgt    5   96005.201 ± 6688.985  ns/op
SHAKE128Benchmark.digest         avgt    5   94778.709 ± 1422.355  ns/op
TupleHash128Benchmark.digest     avgt    5  130159.337 ± 1640.773  ns/op

wasmJs summary:
Benchmark                        Mode  Cnt      Score     Error  Units
M55Benchmark.digest              avgt    5    769.495 ±  29.527  ns/op
SHA1Benchmark.digest             avgt    5   1101.133 ±   3.267  ns/op
SHA256Benchmark.digest           avgt    5   1346.973 ±   1.838  ns/op
SHA512Benchmark.digest           avgt    5   2485.142 ±   7.872  ns/op
ParallelHash128Benchmark.digest  avgt    5  13657.991 ± 881.285  ns/op
SHA3_256Benchmark.digest         avgt    5   2867.188 ±   2.892  ns/op
SHAKE128Benchmark.digest         avgt    5   3048.493 ±  18.849  ns/op
TupleHash128Benchmark.digest     avgt    5   5039.355 ±   3.835  ns/op

nativeHost summary:
Benchmark                        Mode  Cnt      Score     Error  Units
M55Benchmark.digest              avgt    5    497.001 ±  17.779  ns/op
ParallelHash128Benchmark.digest  avgt    5  12537.511 ± 127.138  ns/op
SHA1Benchmark.digest             avgt    5    852.182 ±  13.602  ns/op
SHA256Benchmark.digest           avgt    5   1153.962 ±   6.552  ns/op
SHA3_256Benchmark.digest         avgt    5   3267.204 ±  11.899  ns/op
SHA512Benchmark.digest           avgt    5   1567.941 ±   2.090  ns/op
SHAKE128Benchmark.digest         avgt    5   3500.255 ±  22.726  ns/op
TupleHash128Benchmark.digest     avgt    5   5143.159 ±  30.002  ns/op
```

</details>

<details>
    <summary>Linux</summary>

Before:
```
jvm summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5     642.519 ±   20.925  ns/op
ParallelHash128Benchmark.digest  avgt    5  122080.533 ± 2430.891  ns/op
SHA1Benchmark.digest             avgt    5     680.163 ±   13.619  ns/op
SHA256Benchmark.digest           avgt    5     997.888 ±    5.229  ns/op
SHA3_256Benchmark.digest         avgt    5   36656.908 ±  222.357  ns/op
SHA512Benchmark.digest           avgt    5    1332.470 ±   32.989  ns/op
SHAKE128Benchmark.digest         avgt    5   38390.227 ±  607.338  ns/op
TupleHash128Benchmark.digest     avgt    5   51467.353 ± 1146.[254](https://github.com/KotlinCrypto/hash/actions/runs/12413539131/job/34655644725#step:5:255)  ns/op

js summary:
Benchmark                        Mode  Cnt       Score     Error  Units
M55Benchmark.digest              avgt    5    1862.932 ±  19.021  ns/op
SHA1Benchmark.digest             avgt    5    2139.167 ±   5.393  ns/op
SHA256Benchmark.digest           avgt    5    2577.092 ±   8.650  ns/op
SHA512Benchmark.digest           avgt    5  159692.532 ± 617.359  ns/op
ParallelHash128Benchmark.digest  avgt    5  446908.019 ± 888.664  ns/op
SHA3_256Benchmark.digest         avgt    5  144972.518 ± 634.094  ns/op
SHAKE128Benchmark.digest         avgt    5  145605.971 ± 330.595  ns/op
TupleHash128Benchmark.digest     avgt    5  196839.384 ± 159.965  ns/op

wasmJs summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5    1170.218 ±    2.418  ns/op
SHA1Benchmark.digest             avgt    5    1473.252 ±    3.736  ns/op
SHA256Benchmark.digest           avgt    5    1844.660 ±    9.225  ns/op
SHA512Benchmark.digest           avgt    5    3915.873 ±   16.026  ns/op
ParallelHash128Benchmark.digest  avgt    5  251135.283 ± 1405.252  ns/op
SHA3_256Benchmark.digest         avgt    5   81370.763 ±  117.885  ns/op
SHAKE128Benchmark.digest         avgt    5   81583.087 ±   65.713  ns/op
TupleHash128Benchmark.digest     avgt    5  109096.025 ±  366.819  ns/op

nativeHost summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5     739.297 ±    1.771  ns/op
ParallelHash128Benchmark.digest  avgt    5  536791.078 ± 3963.466  ns/op
SHA1Benchmark.digest             avgt    5     977.585 ±    3.404  ns/op
SHA256Benchmark.digest           avgt    5    1261.492 ±    1.455  ns/op
SHA3_256Benchmark.digest         avgt    5  178200.637 ± 1594.776  ns/op
SHA512Benchmark.digest           avgt    5    1688.339 ±    7.800  ns/op
SHAKE128Benchmark.digest         avgt    5  176545.358 ±  879.950  ns/op
TupleHash128Benchmark.digest     avgt    5  244296.585 ± 1538.095  ns/op
```

After:
```
jvm summary:
Benchmark                        Mode  Cnt      Score     Error  Units
M55Benchmark.digest              avgt    5    611.142 ±  17.012  ns/op
ParallelHash128Benchmark.digest  avgt    5  83476.163 ± 441.402  ns/op
SHA1Benchmark.digest             avgt    5    667.322 ±  16.696  ns/op
SHA256Benchmark.digest           avgt    5   1001.189 ±   7.015  ns/op
SHA3_256Benchmark.digest         avgt    5  26940.596 ± 414.664  ns/op
SHA512Benchmark.digest           avgt    5   1318.751 ±   7.989  ns/op
SHAKE128Benchmark.digest         avgt    5  26737.160 ± 391.069  ns/op
TupleHash128Benchmark.digest     avgt    5  36744.509 ± 458.854  ns/op

js summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5    1884.392 ±    9.625  ns/op
SHA1Benchmark.digest             avgt    5    2223.260 ±    4.599  ns/op
SHA256Benchmark.digest           avgt    5    2672.700 ±    9.460  ns/op
SHA512Benchmark.digest           avgt    5  172120.994 ±  [321](https://github.com/KotlinCrypto/hash/actions/runs/12415706745/job/34662681898?pr=70#step:5:322).879  ns/op
ParallelHash128Benchmark.digest  avgt    5  372871.278 ± 2151.439  ns/op
SHA3_256Benchmark.digest         avgt    5  118554.537 ±  507.883  ns/op
SHAKE128Benchmark.digest         avgt    5  118572.939 ±  334.750  ns/op
TupleHash128Benchmark.digest     avgt    5  162515.505 ±  874.846  ns/op

wasmJs summary:
Benchmark                        Mode  Cnt      Score    Error  Units
M55Benchmark.digest              avgt    5   1152.709 ±  5.674  ns/op
SHA1Benchmark.digest             avgt    5   1463.477 ±  1.548  ns/op
SHA256Benchmark.digest           avgt    5   1771.947 ±  8.167  ns/op
SHA512Benchmark.digest           avgt    5   3938.267 ± 13.947  ns/op
ParallelHash128Benchmark.digest  avgt    5  23571.489 ± 53.127  ns/op
SHA3_256Benchmark.digest         avgt    5   5217.845 ± 17.623  ns/op
SHAKE128Benchmark.digest         avgt    5   5486.102 ± 16.648  ns/op
TupleHash128Benchmark.digest     avgt    5   9471.676 ±  9.882  ns/op

nativeHost summary:
Benchmark                        Mode  Cnt      Score    Error  Units
M55Benchmark.digest              avgt    5    723.059 ±  1.624  ns/op
ParallelHash128Benchmark.digest  avgt    5  20090.627 ± 62.749  ns/op
SHA1Benchmark.digest             avgt    5    955.153 ±  5.283  ns/op
SHA256Benchmark.digest           avgt    5   1251.838 ±  3.969  ns/op
SHA3_256Benchmark.digest         avgt    5   5577.383 ± 29.794  ns/op
SHA512Benchmark.digest           avgt    5   1713.115 ±  5.442  ns/op
SHAKE128Benchmark.digest         avgt    5   6185.740 ± 14.282  ns/op
TupleHash128Benchmark.digest     avgt    5   8618.417 ± 31.880  ns/op
```

</details>

<details>
    <summary>Windows</summary>

Before:
```
jvm summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5     630.902 �   19.714  ns/op
ParallelHash128Benchmark.digest  avgt    5  129542.961 � 1740.712  ns/op
SHA1Benchmark.digest             avgt    5     683.716 �    7.729  ns/op
SHA256Benchmark.digest           avgt    5     996.326 �    7.695  ns/op
SHA3_256Benchmark.digest         avgt    5   37661.863 �  363.563  ns/op
SHA512Benchmark.digest           avgt    5    1352.006 �   12.242  ns/op
SHAKE128Benchmark.digest         avgt    5   38776.227 �  598.766  ns/op
TupleHash128Benchmark.digest     avgt    5   52009.754 �  425.279  ns/op

js summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5    1972.986 ±    9.123  ns/op
SHA1Benchmark.digest             avgt    5    2350.359 ±   28.809  ns/op
SHA256Benchmark.digest           avgt    5    2723.996 ±   12.557  ns/op
SHA512Benchmark.digest           avgt    5  162137.434 ±  265.658  ns/op
ParallelHash128Benchmark.digest  avgt    5  454190.909 ±  555.167  ns/op
SHA3_256Benchmark.digest         avgt    5  148350.896 ± 2380.472  ns/op
SHAKE128Benchmark.digest         avgt    5  147843.695 ±  861.548  ns/op
TupleHash128Benchmark.digest     avgt    5  201235.981 ±  294.049  ns/op

wasmJs summary:
Benchmark                        Mode  Cnt       Score     Error  Units
M55Benchmark.digest              avgt    5    1171.117 ±   1.593  ns/op
SHA1Benchmark.digest             avgt    5    1459.902 ±   2.056  ns/op
SHA256Benchmark.digest           avgt    5    1862.436 ±   3.390  ns/op
SHA512Benchmark.digest           avgt    5    3898.071 ±   8.361  ns/op
ParallelHash128Benchmark.digest  avgt    5  274195.575 ± 917.260  ns/op
SHA3_256Benchmark.digest         avgt    5   88[323](https://github.com/KotlinCrypto/hash/actions/runs/12413539131/job/34655647230#step:5:324).945 ± 171.779  ns/op
SHAKE128Benchmark.digest         avgt    5   88661.363 ± 173.640  ns/op
TupleHash128Benchmark.digest     avgt    5  121150.660 ± 179.841  ns/op

nativeHost summary:
Benchmark                        Mode  Cnt        Score      Error  Units
M55Benchmark.digest              avgt    5      886.604 ±    6.537  ns/op
ParallelHash128Benchmark.digest  avgt    5  1058896.476 ± 4133.226  ns/op
SHA1Benchmark.digest             avgt    5     1113.901 ±    1.980  ns/op
SHA256Benchmark.digest           avgt    5     1406.607 ±    7.316  ns/op
SHA3_256Benchmark.digest         avgt    5   334202.937 ± 1149.855  ns/op
SHA512Benchmark.digest           avgt    5     1842.922 ±   11.608  ns/op
SHAKE128Benchmark.digest         avgt    5   332748.464 ± 1533.510  ns/op
TupleHash128Benchmark.digest     avgt    5   456360.866 ± 3647.229  ns/op
```

After:
```
jvm summary:
Benchmark                        Mode  Cnt      Score      Error  Units
M55Benchmark.digest              avgt    5    588.755 �    1.103  ns/op
ParallelHash128Benchmark.digest  avgt    5  84892.300 � 5490.810  ns/op
SHA1Benchmark.digest             avgt    5    667.711 �   11.003  ns/op
SHA256Benchmark.digest           avgt    5    992.980 �    7.423  ns/op
SHA3_256Benchmark.digest         avgt    5  27081.795 �  130.478  ns/op
SHA512Benchmark.digest           avgt    5   1327.545 �    5.600  ns/op
SHAKE128Benchmark.digest         avgt    5  26724.086 �  166.448  ns/op
TupleHash128Benchmark.digest     avgt    5  37109.222 �  316.020  ns/op

js summary:
Benchmark                        Mode  Cnt       Score      Error  Units
M55Benchmark.digest              avgt    5    2059.892 ±    7.832  ns/op
SHA1Benchmark.digest             avgt    5    2398.781 ±    7.474  ns/op
SHA256Benchmark.digest           avgt    5    2762.966 ±    8.337  ns/op
SHA512Benchmark.digest           avgt    5  160686.520 ±  381.907  ns/op
ParallelHash128Benchmark.digest  avgt    5  377339.805 ± 3352.953  ns/op
SHA3_256Benchmark.digest         avgt    5  120353.889 ± 1285.098  ns/op
SHAKE128Benchmark.digest         avgt    5  120118.581 ±  419.746  ns/op
TupleHash128Benchmark.digest     avgt    5  165877.317 ±  572.520  ns/op

wasmJs summary:
Benchmark                        Mode  Cnt      Score    Error  Units
M55Benchmark.digest              avgt    5   1095.893 ±  2.661  ns/op
SHA1Benchmark.digest             avgt    5   1385.469 ±  1.655  ns/op
SHA256Benchmark.digest           avgt    5   1858.887 ±  1.913  ns/op
SHA512Benchmark.digest           avgt    5   3945.013 ±  7.865  ns/op
ParallelHash128Benchmark.digest  avgt    5  23459.[322](https://github.com/KotlinCrypto/hash/actions/runs/12415706745/job/34662684417?pr=70#step:5:323) ± 61.396  ns/op
SHA3_256Benchmark.digest         avgt    5   5308.779 ± 18.680  ns/op
SHAKE128Benchmark.digest         avgt    5   5626.332 ± 23.821  ns/op
TupleHash128Benchmark.digest     avgt    5   9392.792 ± 14.333  ns/op

nativeHost summary:
Benchmark                        Mode  Cnt      Score     Error  Units
M55Benchmark.digest              avgt    5    874.315 ±   1.259  ns/op
ParallelHash128Benchmark.digest  avgt    5  35579.086 ± 119.435  ns/op
SHA1Benchmark.digest             avgt    5   1110.185 ±   2.149  ns/op
SHA256Benchmark.digest           avgt    5   1529.560 ±   2.443  ns/op
SHA3_256Benchmark.digest         avgt    5   9424.830 ±  55.426  ns/op
SHA512Benchmark.digest           avgt    5   1832.014 ±  10.781  ns/op
SHAKE128Benchmark.digest         avgt    5  10119.581 ±  30.523  ns/op
TupleHash128Benchmark.digest     avgt    5  13975.015 ±  77.548  ns/op
```

</details>